[0.1.x] updated : intBase tests passed
git-svn-id: https://pykd.svn.codeplex.com/svn@71744 9b283d60-5439-405e-af05-b73fd8c4d996

+++ test/scripts/typedvar.py
@@ -52,14 +52,14 @@ class TypedVarTest( unittest.TestCase ):
         self.assertEqual( 0, tv.m_arrayField[0] )
         self.assertEqual( 2, tv.m_arrayField[1] )
         self.assertEqual( 3, tv.m_noArrayField )
-        self.assertNotEqual( -1, long(tv.m_arrayField[0]) )
+        self.assertNotEqual( -1, tv.m_arrayField[0] )
         self.assertNotEqual( 0, tv.m_noArrayField )    
         
     def testGlobalVar(self):
         self.assertEqual( 4, target.module.typedVar( "g_ulongValue" ) )
         self.assertEqual( 0x80000000, target.module.typedVar( "ulongArray" )[3] )
         self.assertEqual( 0x8000000000000000, target.module.typedVar( "ulonglongArray" )[3] )
-        self.assertEqual( -100000,  int(target.module.typedVar( "longArray" )[3]) )        
-        self.assertEqual( -10000000000, long(target.module.typedVar( "longlongArray" )[4]) )
+        self.assertEqual( -100000,  target.module.typedVar( "longArray" )[3])
+        self.assertEqual( -10000000000, target.module.typedVar( "longlongArray" )[4])
         self.assertEqual( target.module.g_structTest, target.module.typedVar( "g_structTestPtr" ) )
